+++ src/components/elements/TextElement/index.ts
@@ -1,2 +1,3 @@
 export { TextElementRenderer } from "./TextElement";
 export type { TextElementRendererProps } from "./TextElement";
+export { PRESENTATION_EDITOR_ACTIONS, normalizeSlideMarkdown } from "./editor-preset";